use lneto for ap example
